Make a workaround for skinnywar/Wildfly
Fixes #7805
